fix: 의뢰비, 성공보수비용 Long 타입으로 수정

+++ src/main/java/com/avg/lawsuitmanagement/lawsuit/controller/form/InsertLawsuitForm.java
@@ -10,8 +10,8 @@ public class InsertLawsuitForm {
     private String lawsuitType;
     private String name;
     private int courtId;   // 담당법원 -> int로 받을지 string으로 받을지?
-    private int commissionFee; // 의뢰비
-    private int contingentFee; // 성공보수비용
+    private Long commissionFee; // 의뢰비
+    private Long contingentFee; // 성공보수비용
     private String lawsuitNum;
     private List<Long> memberId;
     private List<Long> clientId;
