+++ src/main/java/com/avg/lawsuitmanagement/lawsuit/controller/form/UpdateLawsuitInfoForm.java
@@ -12,8 +12,8 @@ public class UpdateLawsuitInfoForm {
     private String lawsuitType;
     private String name;
     private int courtId;
-    private int commissionFee;
-    private int contingentFee;
+    private Long commissionFee;
+    private Long contingentFee;
     private String lawsuitStatus;
     private String lawsuitNum;
     private String result;
